requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.35
pydantic==2.9.0
yfinance==0.2.43
nsetools==1.0.12
ta==0.11.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
tensorflow-cpu==2.17.0
prophet==1.1.6
websockets==13.0
aiohttp==3.10.5
python-dotenv==1.0.1
feedparser==6.0.11
smartapi-python==1.3.6
pyotp==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
xgboost==2.1.1
curl_cffi>=0.7.0
